requirements.txt
Django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
Pillow
celery>=5.3
redis>=4.5
